+++ lib/rubocop/cop/style/trivial_accessors.rb
@@ -78,7 +78,8 @@ module RuboCop
         end
 
         def looks_like_trivial_writer?(args, body)
-          args.children.size == 1 && args.children[0].type != :restarg &&
+          args.children.size == 1 &&
+            ![:restarg, :blockarg].include?(args.children[0].type) &&
             body && body.type == :ivasgn &&
             body.children[1] && body.children[1].type == :lvar
         end